requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pillow==10.3.0
psycopg2-binary==2.9.9
pyotp==2.9.0
pypng==0.20220715.0
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.31.0
telebot==0.0.5
typing_extensions==4.11.0
urllib3==2.2.1
